Refactor tests - WIP
